Move `func*` files into `aws-func`
As per discussion and agreement, moving the aws specific function implementation details into a separate folder that I have decided to call `aws-func`.
In that folder, I renamed the "load everything" file to index.js to clean up requiring it.
I then renamed each of the files to better represent their purpose.  Previously they were func____.js to differentiate them from the not faas provider specific files.
Lots of reference updates.
