+++ lib/faas/aws-func/definitions.js
