+++ lib/faas/aws-func/handle.js
@@ -8,8 +8,8 @@ const promisify = require('util-promisify')
 const {
   MAX_TIMEOUT_BUFFER_IN_MILLISECONDS: maxTimeoutBuffer,
   MERGE_FIELD: mergeFileField,
-} = require('./funcDef')
-const valid = require('./funcValid')
+} = require('./definitions')
+const valid = require('./valid')
 
 const readFileAsync = promisify(fs.readFile)
