+++ lib/faas/aws-func/index.js
@@ -1,0 +1,11 @@
+const def = require('./definitions')
+const exec = require('./invoke')
+const handle = require('./handle')
+const valid = require('./valid')
+
+module.exports = {
+  def,
+  exec,
+  handle,
+  valid,
+}
