+++ lib/faas/aws-func/invoke.js
@@ -1,6 +1,6 @@
 const aws = require('aws-sdk') // eslint-disable-line import/no-extraneous-dependencies
 
-const valid = require('./funcValid')
+const valid = require('./valid')
 
 const lambda = new aws.Lambda({ maxRetries: 0 })
