+++ lib/faas/aws-func/valid.js
@@ -1,4 +1,4 @@
-const def = require('./funcDef')
+const def = require('./definitions')
 
 const impl = {
   /**
