--- lib/faas/func.js
@@ -1,11 +1,0 @@
-const def = require('./funcDef')
-const exec = require('./funcExec')
-const handle = require('./funcHandle')
-const valid = require('./funcValid')
-
-module.exports = {
-  def,
-  exec,
-  handle,
-  valid,
-}
