+++ lib/faas/handler.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-underscore-dangle */
 
 const alert = require('./alert')
-const func = require('./func')
+const func = require('./aws-func')
 const task = require('./task')
 
 const impl = {
