+++ lib/index.js
@@ -9,7 +9,7 @@ const stdin = require('get-stdin')
 const url = require('url')
 const yaml = require('js-yaml')
 
-const func = require('./faas/func')
+const func = require('./faas/aws-func')
 const task = require('./faas/task')
 const npm = require('./npm')
 const Serverless = require('./serverless-fx')
@@ -21,7 +21,7 @@ const constants = {
   ServerlessFiles: [
     'alert.js',
     'handler.js',
-    'func.js', 'funcDef.js', 'funcExec.js', 'funcHandle.js', 'funcValid.js',
+    'aws-func/index.js', 'aws-func/definitions.js', 'aws-func/invoke.js', 'aws-func/handle.js', 'aws-func/valid.js',
     'package.json', 'package-lock.json',
     'serverless.yml',
     'task.js', 'taskDef.js', 'taskExec.js', 'taskPlan.js', 'taskResult.js', 'taskValid.js',
