+++ tests/lib/faas/funcValid.spec.js
@@ -4,7 +4,7 @@ const path = require('path')
 const { expect } = chai
 
 // eslint-disable-next-line import/no-dynamic-require
-const func = require(path.join('..', '..', '..', 'lib', 'faas', 'func.js'))
+const func = require(path.join('..', '..', '..', 'lib', 'faas', 'aws-func'))
 
 const tagScript = () => ({
   config: {
@@ -16,7 +16,7 @@ const tagScript = () => ({
 
 let script
 
-describe('./lib/faas/funcValid.js', () => {
+describe('./lib/faas/aws-func/valid.js', () => {
   describe(':impl', () => {
     describe('#validScript', () => {
       /* eslint-disable no-underscore-dangle */
